rename fb cache and write docstring
